Get rid of reconnect_soon
