+++ aiohomekit/controller/abstract.py
@@ -228,13 +228,6 @@ class AbstractPairing(metaclass=ABCMeta):
     async def unsubscribe(self, characteristics: Iterable[tuple[int, int]]) -> None:
         self.subscriptions.difference_update(characteristics)
 
-    async def reconnect_soon(self) -> None:
-        """
-        Notify the pairing that we have noticed a network change that means its connection maybe stale.
-
-        This will be removed in a future release.
-        """
-
     def dispatcher_availability_changed(
         self, callback: Callable[[bool], None]
     ) -> Callable[[], None]:
